[update] PupulationChartのロジック分離
ロジック部分をカスタムフック化
